1.2.0   moved pyrometer to separate pyro class and forced port close after each interaction
