+++ app_control.py
@@ -6,7 +6,7 @@ import random
 import json
 from datetime import datetime
 
-VERSION = '1.1.13'
+VERSION = '1.2.0'
 
 def initialise():
     """Setup the settings structure with default values"""
